requirements.txt
rich>=13.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
aiohttp>=3.8.0
aiosqlite>=0.19.0
numpy>=1.24.0
torch>=2.0.0
transformers>=4.30.0
psutil>=5.9.0